Rename enum files to enumType
