+++ BabelWiresLib/Types/Enum/addBlankToEnum.cpp
@@ -9,7 +9,7 @@
 
 #include <BabelWiresLib/TypeSystem/typeSystem.hpp>
 #include <BabelWiresLib/TypeSystem/typeSystemException.hpp>
-#include <BabelWiresLib/Types/Enum/enum.hpp>
+#include <BabelWiresLib/Types/Enum/enumType.hpp>
 
 babelwires::ShortId babelwires::AddBlankToEnum::getBlankValue() {
     return BW_SHORT_ID("____", "____", "9e715cdc-399a-48b7-bc04-ad98e9e900d9");
