+++ BabelWiresLib/Types/Enum/addBlankToEnum.hpp
@@ -8,7 +8,7 @@
 #pragma once
 
 #include <BabelWiresLib/TypeSystem/typeConstructor.hpp>
-#include <BabelWiresLib/Types/Enum/enum.hpp>
+#include <BabelWiresLib/Types/Enum/enumType.hpp>
 
 namespace babelwires {
     /// A unary type constructor which adds a blank value to an existing enum.
