+++ BabelWiresLib/Types/Enum/enumWithCppEnum.hpp
@@ -7,7 +7,7 @@
  **/
 #pragma once
 
-#include <BabelWiresLib/Types/Enum/enum.hpp>
+#include <BabelWiresLib/Types/Enum/enumType.hpp>
 #include <BabelWiresLib/Types/Enum/enumFeature.hpp>
 
 #include <Common/Identifiers/registeredIdentifier.hpp>
