+++ BabelWiresLib/Types/Map/Helpers/enumValueAdapters.hpp
@@ -7,7 +7,7 @@
  **/
 #pragma once
 
-#include <BabelWiresLib/Types/Enum/enum.hpp>
+#include <BabelWiresLib/Types/Enum/enumType.hpp>
 #include <BabelWiresLib/Types/Enum/enumValue.hpp>
 
 namespace babelwires {
