+++ BabelWiresLib/Types/Map/Helpers/mapApplicatorFallbackHelper.hpp
@@ -7,7 +7,7 @@
  **/
 #pragma once
 
-#include <BabelWiresLib/Types/Enum/enum.hpp>
+#include <BabelWiresLib/Types/Enum/enumType.hpp>
 #include <BabelWiresLib/Types/Map/Helpers/valueAdapter.hpp>
 #include <BabelWiresLib/Types/Map/MapEntries/allToOneFallbackMapEntryData.hpp>
 #include <BabelWiresLib/Types/Map/MapEntries/oneToOneMapEntryData.hpp>
