+++ BabelWiresQtUi/ComplexValueEditors/MapEditor/typeWidget.cpp
@@ -3,7 +3,7 @@
 #include <BabelWiresQtUi/uiProjectContext.hpp>
 
 #include <BabelWiresLib/TypeSystem/typeSystem.hpp>
-#include <BabelWiresLib/Types/Enum/enum.hpp>
+#include <BabelWiresLib/Types/Enum/enumType.hpp>
 
 #include <Common/Identifiers/identifierRegistry.hpp>
