+++ BabelWiresQtUi/ValueModels/enumValueModel.cpp
@@ -9,7 +9,7 @@
 
 #include <BabelWiresQtUi/ValueEditors/dropDownValueEditor.hpp>
 
-#include <BabelWiresLib/Types/Enum/enum.hpp>
+#include <BabelWiresLib/Types/Enum/enumType.hpp>
 #include <BabelWiresLib/Types/Enum/enumValue.hpp>
 
 #include <Common/Identifiers/identifierRegistry.hpp>
